Release (#76)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## photo-id@0.1.4

### Patch Changes

-   1728568: Use Vitest
-   28e2ee9: Fix async directory checks
-   51e07c1: Update import linting
-   181aae6: Update prettier config
-   98a3def: Refactor IPC

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>