Release (#84)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## photo-id@0.2.0

### Minor Changes

- 3065042: Selected photos in stacks are now retained when opening a
project

### Patch Changes

-   2aba323: Double-clicking an image now opens it in the editor

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>