Release (#88)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## photo-id@0.2.1

### Patch Changes

-   2df6288: Fixed exporting "L" and "R" labels in filenames

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>