Release (#90)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## photo-id@0.2.2

### Patch Changes

-   fd2d9ff: Optimise project initialisation
- ff6c258: Improve performance when using the app for long periods of
time
-   9689894: Fixed bug when removing the penultimate photo from a stack

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>