Release (#101)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## photo-id@0.2.4

### Patch Changes

-   1150977: Add React and MobX developer tools
- 18c69ee: Show error when attempting to open a recent project that no
longer exists
-   620bcc3: Add initial documentation
-   a3262f7: Add Sentry error tracking

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>