Release (#127)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## photo-id@0.3.0

### Minor Changes

-   0326d8f: Update app logo

### Patch Changes

- e8089aa: App now focuses when a new project has finished initialising
-   715c4bc: Show default window title when closing a project
-   e321e06: Add changelog button
- 5353878: Fix closing edit windows when closing a project multiple
times
- 0ba46f6: Fix error when closing projects after opening and closing
edit windows
-   b5fdf6c: Closing a project now closes all open edit windows
-   464d160: Fix thumbnails not updating when editing images
-   23e780d: Limit photo editor navigation

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>